Plane: return pilot descent speed as float, not uint16_t
Since the SI conversion Q_PILOT_SPD_UP and Q_PILOT_SPD_DN are AP_Float
metres per second (convert_centi_parameter), but
get_pilot_velocity_z_max_dn_m() kept its uint16_t return type. Any
fractional value is truncated to whole m/s: with Q_PILOT_SPD_UP 0.75 and
Q_PILOT_SPD_DN 0 (the default fallback) the descent limit handed to the
position controller is 0 m/s, so full-down stick in QHOVER/QLOITER
commands no descent at all; Q_PILOT_SPD_DN 1.5 becomes 1.

Return float and use fabsf(); fix the stale "cm/s" comment.

Observed on a quadplane in SITL (PSCD.DVD 0.00 with the stick at the
bottom, 0.59 near the top; setting Q_PILOT_SPD_DN 1 restores descent).

+++ ArduPlane/quadplane.h
@@ -152,7 +152,7 @@ public:
     // is throttle controlled landing descent active?
     bool thr_ctrl_land;
 
-    uint16_t get_pilot_velocity_z_max_dn_m() const;
+    float get_pilot_velocity_z_max_dn_m() const;
     
     struct PACKED log_QControl_Tuning {
         LOG_PACKET_HEADER;
